ch348: Implement status report / interrupt handling
We have to wait for the R_II_B2 status / interrupt after transmitting
one URB, otherwise we (can) overflow the HW transmit buffer and data
gets corrupted. Doing so is a bit tricky as usb_serial_generic_write()
wants to process the whole buffer without giving the driver a chance to
wait in between USB transfers. Only passing N bytes that can be
transferred in one URB to usb_serial_generic_write() means that we can
wait afterwards and let TTY core call .write() again with the remaining
buffer and length (if any).

This fixes TX buffer overflows when sending more than 509 bytes at a
time (but especially larger transfers of 1024 or 2024 bytes are often
seen to be affected).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>