ch348: use a mutex to protect against concurrent writes
The vendor driver prevents concurrent writes with a spin lock. Prevent
them in the upstream driver as well but use a mutex as the code within
the critical section may sleep. This fixes "Failed to wait for TX buffer
flush" errors for slow baud rates when reading and writing two CH348
ports that are connected together.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>